Include MarketOracle Entity
Initial Proposal to include a MarketOracle entity. This can be used to determine the conversion rate of a market's tokens as per how the protocol also determines the conversion.

Also, I added (immutable: true) to event level data since this should not be updated after `.save()` which should save some db space (I believe the block_range column in the table is omitted when immutable: true).